updated calculator

+++ Calculator.py
@@ -1,4 +1,4 @@
-num1 = int(input("enter the number you want to add too:"))
-num2 = int(input("enter the number you want to add too it:"))
+num1 = float(input("enter the number you want to add too:"))
+num2 = float(input("enter the number you want to add too it:"))
 num3 = num1 + num2
 print(num3)
